requirements.txt
flake8==7.0.0
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.2.0
PyYAML==6.0.1
